+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - Add `Animation.stop()` to stop an animation without clearing it. [(#95)](https://github.com/ajalt/mordant/issues/95)
 - Animations now support resuming after a call to `stop` or `clear`. [(#94)](https://github.com/ajalt/mordant/issues/94)
+- `TextStyles.reset`, `TextStyles.resetForeground`, and `TextStyles.resetBackground` to clear existing styles.
 
 ### Fixed
 - Fix `TerminalInfo.interactive` not including `outputInteractive`
@@ -11,6 +12,8 @@
 
 ### Changed
 - verticalLayout{} now defaults `TextAlign.NONE`, meaning it won't add any trailing whitespace to lines. You can return to the old behavior with `align = TextAlign.LEFT`.
+- All boolean fields on `TextStyle` are now nullable. A null field indicates no change to the previs value when adding or nesting style.
+- When nesting styles, the outer style will now override inner styles at the start of a string. (e.g. `red(blue("x")) == red("x")`) 
 
 ## 2.0.0-beta12
 ### Fixed